EM-1933: circular dependency
