+++ package.json
@@ -4,11 +4,12 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "embeddable:build": "embeddable build",
+    "embeddable:build": "npm run check-circular-dependencies && embeddable build",
     "embeddable:push": "embeddable push",
     "embeddable:login": "embeddable login",
-    "embeddable:dev": "cross-env NODE_OPTIONS='--max-old-space-size=4096' embeddable dev",
+    "embeddable:dev": "npm run check-circular-dependencies && cross-env NODE_OPTIONS='--max-old-space-size=4096' embeddable dev",
     "embeddable:update": "npx npm-check-updates -u \"/embeddable/\"",
+    "check-circular-dependencies": "npx madge --circular src --extensions ts,tsx",
     "check-types": "tsc --noemit",
     "ct": "tsc --noEmit",
     "cube:playground": "docker-compose -f ./src/cube-playground.yml up",
@@ -84,6 +85,7 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.14",
+    "madge": "^8.0.0",
     "postcss": "^8.4.48",
     "prettier": "3.3.3",
     "tailwindcss": "^3.3.2",
