+++ src/components/vanilla/DownloadMenu.tsx
@@ -6,7 +6,7 @@ import IconDownloadPNG from '../icons/DownloadPNG';
 import IconVerticalEllipsis from '../icons/VerticalEllipsis';
 import downloadAsCSV from '../util/downloadAsCSV';
 import downloadAsPNG from '../util/downloadAsPNG';
-import { ContainerProps } from './Container';
+import { ContainerProps } from './Container.types';
 
 interface CSVProps extends ContainerProps {
   results?: DataResponse | DataResponse[];
